Remove 32bit version of tryteValuesToBytes
